requirements.txt
altgraph==0.17.3
colorama==0.4.6
packaging==23.1
pefile==2023.2.7
Pillow==9.5.0
pyinstaller==5.11.0
pyinstaller-hooks-contrib==2023.3
pyproject_hooks==1.0.0
pywin32-ctypes==0.2.0
tomli==2.0.1
